deluxe_core

Macro only_one

macro_rules! only_one {
    ($prefix:expr, $errors:expr $(, $fields:tt)* $(,)?) => { ... };
}

Appends an error if more than one given attribute is present, automatically casting and stringifying field names.

Convenience macro for only_one. The first argument is a &str prefix to prepend onto the names in the event of an error. The second argument is an &Errors to add any errors into.

The rest of the arguments are either identifiers or 2-element tuples. Identifiers must be an Option<T> and tuples must be (&str, Option<&T>), where &T is castable to &dyn syn::spanned::Spanned. Identifiers will automatically stringify the identifier name to use in the error message, and so should used when the field name is the same as the variable name. The tuple form can be used to specify a different field name from the name of the variable. The identifier form is intended to make it easier to do validations by destructuring a data type after it has been parsed.

Example

#[derive(Default)]
struct MyStruct {
    my_bool: Option<syn::LitBool>,
    path: Option<syn::Path>,
    // #[deluxe(rename = "ty")]
    field3: Option<syn::Type>,
}
let errors = deluxe_core::Errors::new();
let s = MyStruct::default();

// ... parsing omitted ...

let MyStruct { my_bool, path, field3 } = &s;
// only one of `my_bool`, `path`, `ty` is allowed
deluxe_core::only_one!("", &errors, my_bool, path, ("ty", field3.as_ref()));