deluxe_core

Macro parse_named_meta_item_with

macro_rules! parse_named_meta_item_with {
    ($input:expr, $name_span:expr, $($path:tt)* $(,)?) => { ... };
}

Parses a ParseMetaItem following a name, using a custom parse module.

Equivalent to parse_named_meta_item_with, but the functions are taken from a module path. Can be used to implement additional parsing strategies for types that already implement ParseMetaItem, or for generic types. Some common parsers are available in the with module.

The first argument is a syn::parse::ParseStream, the second argument is the Span of the name tokens, and the third argument is a module path. Modules are required to implement the functions parse_meta_item, parse_meta_item_inline, and parse_meta_item_flag, even if just to return an error. The signatures of these functions should match the equivalent functions in ParseMetaItem.

parse_meta_item_flag implementations can call flag_disallowed_error for a standard error if flags are not supported by the target type.