deluxe_core
Module parse_helpers

Utility functions for parse trait implementations.

Structs

Brace

{…}

Bracket

[…]

Nothing

An empty syntax tree node that consumes no tokens when parsed.

Paren

(…)

SmallString

A string type that can be stored on the stack or on the heap, and can also represent borrowed strings.

Enums

FieldStatus

Temporary container for storing parsing state for a single field.

NamedParse

The type of a named meta field.

Traits

ParseDelimited

Helpers for parsing tokens inside a delimiter.

Functions

check_unknown_attribute

Appends an error if a field is not included in a list of fields.

disallow_paths

Appends errors if any keys are found in a list of paths.

first_path_name

Gets the first string in a list of path names and spans.

first_span

Gets the first Span in a list of path names and spans.

flag_disallowed_error

Returns an error with a “unexpected flag” message.

fork_inputs

Forks all streams in a ParseStream slice into a new Vec of owned ParseBuffers.

has_paths

Checks if all keys are contained in a set of paths.

inputs_span

Creates a span encompassing a collection of ParseStreams.

join_path

Concatenates two path strings.

join_paths

Calls join_path on all paths in a slice.

join_prefix

Concatenates two path prefix strings.

key_to_string

Converts a ToKeyString to a SmallString.

missing_field_error

Returns an error with a “missing required field” message.

only_one_variant

Appends a “too many enum variants” error.

parse_any_path

Parses a path allowing any keywords as identifiers, and containing no path arguments.

parse_empty

Runs a parsing function from an empty token stream.

parse_empty_meta_item

Parses a ParseMetaItem inline from an empty token stream.

parse_eof_or_trailing_comma

Parses a stream for end-of-stream, or a single comma.

parse_first

Runs a parsing function on the first stream in a stream list, then ensures the rest of the streams are empty.

parse_named_meta_item

Parses a ParseMetaItem following a name.

parse_named_meta_item_with

Parses a ParseMetaItem following a name, using custom parse functions.

parse_struct

Parses an inline struct with named fields from a stream.

parse_struct_attr_tokens

Parses a collection of ToTokens values into a struct with named fields.

parse_tuple_struct

Parses an inline fixed-length tuple struct from a stream.

path_matches

Checks if a path matches a list of names.

peek_eof_or_trailing_comma

Checks if a stream is empty or contains only a comma.

ref_tokens

Returns an iterator of pairs of TokenStreams and the corresponding path Span from a matched ParseAttributes.

remove_paths

Removes some paths from a path hashmap.

skip_all

Consumes all tokens left in a list of streams.

skip_meta_item

Skips over items in a ParseStream until a comma is reached.

take_tokens

Returns an iterator of TokenStreams and the corresponding path Span from a matched ExtractAttributes.

try_parse_named_meta_item

Attempts to begin parsing a meta item following a name, returning a value describing what type of named meta item this is.

unknown_error

Returns an error with an “unknown field” message.

variant_required

Appends a “enum variant expected” error message.