deluxe_core::parse_helpers

Enum FieldStatus

pub enum FieldStatus<T> {
    None,
    ParseError,
    Some(T),
}

Temporary container for storing parsing state for a single field.

Variants

None

Field is not present or not parsed yet.

ParseError

Field was present, but had an error during parsing.

Some(T)

Field was found and successfully parsed.

Implementations

impl<T> FieldStatus<T>

pub const fn as_ref(&self) -> FieldStatus<&T>

Converts from &FieldStatus<T> to FieldStatus<&T>.

pub fn and<U>(self, b: FieldStatus<U>) -> FieldStatus<U>

Returns b if the status is Some.

pub fn and_then<U, F: FnOnce(T) -> FieldStatus<U>>(self, f: F) -> FieldStatus<U>

If the status is Some, calls f with the wrapped value and returns the result.

pub fn map<U, F: FnOnce(T) -> U>(self, f: F) -> FieldStatus<U>

Maps a FieldStatus<T> to FieldStatus<U> by applying a function to a contained value.

pub const fn is_none(&self) -> bool

Returns true if the field is a FieldStatus::None value.

pub const fn is_parse_error(&self) -> bool

Returns true if the field is a FieldStatus::ParseError value.

pub const fn is_some(&self) -> bool

Returns true if the field is a FieldStatus::Some value.

pub fn or(self, b: FieldStatus<T>) -> FieldStatus<T>

Returns the status if it contains a value, or if it is None then returns b.

pub fn or_else<F: FnOnce() -> FieldStatus<T>>(self, f: F) -> FieldStatus<T>

Returns the status if it contains a value, or if it is None then calls f and returns the result.

pub fn unwrap(self) -> T

Returns the contained FieldStatus::Some value, consuming the self value.

Panics

Panics if the self value equals FieldStatus::None or FieldStatus::ParseError.

pub fn unwrap_or(self, default: T) -> T

Returns the contained FieldStatus::Some value or a provided default.

pub fn unwrap_or_else<F: FnOnce() -> T>(self, f: F) -> T

Returns the contained FieldStatus::Some value or computes it from a closure.

pub fn flatten<U>(self) -> FieldStatus<U>
where T: Into<FieldStatus<U>>,

Converts from FieldStatus<impl Into<FieldStatus<T>>> to FieldStatus<T>.

pub fn parse_named_item( &mut self, name: &str, input: ParseStream<'_>, span: Span, errors: &Errors, )
where T: ParseMetaItem,

Parses a named item into this status.

pub fn parse_named_item_with<F>( &mut self, name: &str, input: ParseStream<'_>, span: Span, errors: &Errors, func: F, )
where F: FnOnce(ParseStream<'_>, &str, Span) -> Result<T>,

Parses a named item into this status, using a custom parsing function.

pub fn parse_unnamed_item(&mut self, input: ParseStream<'_>, errors: &Errors)
where T: ParseMetaItem,

Parses an unnamed item into this status.

pub fn parse_unnamed_item_with<F>( &mut self, input: ParseStream<'_>, errors: &Errors, func: F, )
where F: FnOnce(ParseStream<'_>, ParseMode) -> Result<T>,

Parses an unnamed item into this status, using a custom parsing function.

pub fn into_option(self) -> Option<T>

Converts the status to an Option.