deluxe_core::parse_helpers

Function parse_named_meta_item

pub fn parse_named_meta_item<T: ParseMetaItem>(
    input: ParseStream<'_>,
    name_span: Span,
) -> Result<T>

Parses a ParseMetaItem following a name.

To be used by any parse implementations that expect named attributes.

If the first token is =, then calls ParseMetaItem::parse_meta_item. If the first token is a parenthesized group, calls ParseMetaItem::parse_meta_item_inline on the contents.

If neither a = or ( is parsed, then no additional tokens are consumed, and ParseMetaItem::parse_meta_item_flag is called with the current stream span.