deluxe_core::parse_helpers

Function parse_struct_attr_tokens

pub fn parse_struct_attr_tokens<T, I, F, R>(inputs: I, func: F) -> Result<R>
where
    T: ToTokens,
    I: IntoIterator<Item = (T, SmallString<'static>, Span)>,
    F: FnOnce(&[ParseBuffer<'_>], &[(SmallString<'static>, Span)]) -> Result<R>,

Parses a collection of ToTokens values into a struct with named fields.

Should be used from ParseAttributes or ExtractAttributes implementations, which should respectively call ref_tokens or take_tokens before passing any token streams into this function.

Each stream will be parsed for a parenthesized group. Then, func will be called with a list of every ParseBuffer inside these groups, and the corresponding Span for the path of that group. An end-of-stream (or trailing comma followed by end-of-stream) will be consumed from each buffer after func returns. Callers should ensure the success path consumes all other tokens in each buffer.

Structs with named fields or tuple structs should be parsed by respectively calling parse_struct or parse_tuple_struct.

Returns a parse error if any of the streams failed to parse. Otherwise, passes through the return value from func.