deluxe_core::parse_helpers

Function parse_tuple_struct

pub fn parse_tuple_struct<'s, F, S>(
    inputs: &[S],
    len: usize,
    func: F,
) -> Result<usize>
where
    S: Borrow<ParseBuffer<'s>>,
    F: FnMut(ParseStream<'_>, &[S], usize) -> Result<()>,

Parses an inline fixed-length tuple struct from a stream.

len is the maximum number of items to parse. Each stream in inputs will be parsed in order, calling func for every field with three arguments: the current stream, a slice containing the current stream and all following streams, and the index of the current field starting from 0. Commas will be parsed between each field.

Returns the number of items that were successfully parsed, or Err on any parsing errors or the first failure of func. A trailing comma will be consumed only if the return value is less than len.