deluxe_core

Struct SpannedValue

pub struct SpannedValue<T> { /* private fields */ }

A wrapper for adding a Span to an arbitrary value.

Implementations are provided for all the parsing traits that simply delegate to T, capturing the inital Span from the ParseStream.

Implementations

impl<T> SpannedValue<T>

pub fn new(value: T) -> Self

Creates a new value wrapping a T, with the span set to Span::call_site.

pub fn with_span(value: T, span: Span) -> Self

Creates a new value wrapping a T, with the span set to span.

pub fn into_inner(value: SpannedValue<T>) -> T

Unwraps a SpannedValue into a T. Note this is an associated function, not a method.