deluxe_core

Trait HasAttributes

pub trait HasAttributes {
    // Required methods
    fn attrs(&self) -> &[Attribute];
    fn attrs_mut(&mut self) -> Result<&mut Vec<Attribute>>;
}

Trait for a syn type containing a list of attributes.

Implementations are provided for all syn types containing a Vec<syn::Attribute>.

Required Methods

fn attrs(&self) -> &[Attribute]

Returns an immutable slice of attributes.

fn attrs_mut(&mut self) -> Result<&mut Vec<Attribute>>

Returns a mutable Vec of attributes.

Returns Err if the type does not support mutable attributes.