deluxe_core

Trait ParseAttributes

pub trait ParseAttributes<'t, T: HasAttributes>: Sized {
    // Required methods
    fn path_matches(path: &Path) -> bool;
    fn parse_attributes(obj: &'t T) -> Result<Self>;
}

Parses a structure out of a syn::Attribute list.

This trait is intended for types that may share a set of matching attributes with other types.

Required Methods

fn path_matches(path: &Path) -> bool

Checks if a given attribute path can be parsed by this type.

fn parse_attributes(obj: &'t T) -> Result<Self>

Iterates the attributes from a T, parsing any that match.

Implementations should use the ref_tokens helper to automatically get references to any matching TokenStreams, and then parse them using parse_struct_attr_tokens.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.