deluxe_core

Trait ParseMetaAppend

pub trait ParseMetaAppend: Sized {
    // Required method
    fn parse_meta_append<'s, S, I, P>(inputs: &[S], paths: I) -> Result<Self>
       where S: Borrow<ParseBuffer<'s>>,
             I: IntoIterator<Item = P>,
             I::IntoIter: Clone,
             P: AsRef<str>;
}

Parses a meta item for a structure with named fields that concatenates all matching items.

Should be implemented on collection types. Implementations are provided for the common collection types in std.

Required Methods

fn parse_meta_append<'s, S, I, P>(inputs: &[S], paths: I) -> Result<Self>
where S: Borrow<ParseBuffer<'s>>, I: IntoIterator<Item = P>, I::IntoIter: Clone, P: AsRef<str>,

Parse the item from a group of inline named contexts.

Fields with names matching any path in paths will be appended. Non-matching fields should be skipped with crate::parse_helpers::skip_meta_item.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.