deluxe_core

Trait ParseMetaFlatNamed

pub trait ParseMetaFlatNamed: Sized {
    const ACCEPTS_ALL: bool = false;

    // Required methods
    fn field_names() -> &'static [&'static str];
    fn parse_meta_flat_named<'s, S: Borrow<ParseBuffer<'s>>>(
        inputs: &[S],
        _mode: ParseMode,
        prefix: &str,
        validate: bool,
    ) -> Result<Self>;
}

Parses a meta item for a structure with named fields.

Automatically implemented by derive(ParseMetaItem) for structures with named fields, and for enums.

A parent structure that wants to flatten items of this type into its own fields should call Self::parse_meta_flat_named.

Provided Associated Constants

const ACCEPTS_ALL: bool = false

A flag noting if this parser will consume all unknown fields.

Should be set to true only for structures containing a #[deluxe(rest)] field.

Required Methods

fn field_names() -> &'static [&'static str]

Returns an array specifying all optional and required fields accepted by this structure.

fn parse_meta_flat_named<'s, S: Borrow<ParseBuffer<'s>>>( inputs: &[S], _mode: ParseMode, prefix: &str, validate: bool, ) -> Result<Self>

Parse the item from a group of inline named contexts.

The streams can consume any number of commas. The streams in inputs should not be considered contiguous, but should imply a trailing comma if one is not present. Implementations should parse each stream in order, continuing to the next stream when the current one is exhausted.

All streams should be parsed to the end, skipping over unknown fields, and consuming a trailing comma if present. The crate::parse_helpers::parse_struct helper should be used for convenience. prefix should be stripped off the names of all fields before pattern matching on them.

If validate is true, then an error should be generated upon encountering any fields not in Self::field_names.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.