deluxe_core

Trait ParseMetaFlatUnnamed

pub trait ParseMetaFlatUnnamed: Sized {
    // Required methods
    fn field_count() -> Option<usize>;
    fn parse_meta_flat_unnamed<'s, S: Borrow<ParseBuffer<'s>>>(
        inputs: &[S],
        _mode: ParseMode,
        index: usize,
    ) -> Result<Self>;
}

Parses a meta item for a structure with unnamed fields.

Automatically implemented by derive(ParseMetaItem) for tuple structures, but can also be manually implemented for array-like types.

A parent structure that wants to flatten items of this type into its own fields should call Self::parse_meta_flat_unnamed.

Required Methods

fn field_count() -> Option<usize>

Returns the number of fields in this structure.

Returns None if it can parse any number of fields.

Currently, this is only useful when generating error messages.

fn parse_meta_flat_unnamed<'s, S: Borrow<ParseBuffer<'s>>>( inputs: &[S], _mode: ParseMode, index: usize, ) -> Result<Self>

Parse the item from a group of inline contexts.

The streams can consume any number of commas. The streams in inputs should not be considered contiguous, but should imply a trailing comma if one is not present. Implementations should parse each stream in order, continuing to the next stream when the current one is exhausted.

If the structure can take any number of fields, it should parse all streams to the end and ensure that a traiing comma is parsed.

If the structure contains a finite number of fields, then trailing comma should not be consumed. Once all fields are parsed, the function should return, without any further modifications to the current stream or any following streams.

index is the starting offset into the currently parsing tuple. It should be used as a base when generating error messages.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.