deluxe_core

Trait ParseMetaItem

pub trait ParseMetaItem: Sized {
    // Required method
    fn parse_meta_item(input: ParseStream<'_>, _mode: ParseMode) -> Result<Self>;

    // Provided methods
    fn parse_meta_item_inline<'s, S: Borrow<ParseBuffer<'s>>>(
        inputs: &[S],
        _mode: ParseMode,
    ) -> Result<Self> { ... }
    fn parse_meta_item_flag(_span: Span) -> Result<Self> { ... }
    fn parse_meta_item_named(
        input: ParseStream<'_>,
        _name: &str,
        span: Span,
    ) -> Result<Self> { ... }
    fn missing_meta_item(name: &str, span: Span) -> Result<Self> { ... }
}

Base trait for parsing a single field out of syn::parse::ParseStream.

This trait is conceptually similar to syn::parse::Parse, but with a few key differences:

It can tell when an item is being parsed from a tuple struct or a struct with named fields.

It can be parsed “inline” by containers that consume the surrounding delimiters.

It can be parsed as a “flag” by containers that allow empty fields.

It can provide custom parsing when it is used with a named attribute.

Implementations are provided for all primitives and standard collections, as well as for tuples, arrays, and for most parseable node structures in syn.

Required Methods

fn parse_meta_item(input: ParseStream<'_>, _mode: ParseMode) -> Result<Self>

Parse the item from the tokens in input.

If the item can contain commas, the whole item should be wrapped in a pair of delimiters. The stream should should not parse any trailing commas or additional tokens past the end of the item.

The _mode argument describes whether this item is being parsed in a named or an unnamed context. This argument will rarely be used. It only should be checked when an item uses Self::parse_meta_item_flag to provide shorthand for a default enum value, and wants to avoid parsing it in a named context. In an unnamed context, both values will still need to be parsed. See the source to the implementation of this trait on Option<T> for an example of when to check _mode.

Provided Methods

fn parse_meta_item_inline<'s, S: Borrow<ParseBuffer<'s>>>( inputs: &[S], _mode: ParseMode, ) -> Result<Self>

Parse the item in an inline context.

The stream can consume any number of commas. Items with a finite length should not consume a trailing comma. See Self::parse_meta_item for a description of the _mode argument.

This function should be implemented for custom data structures, but is not intended to be called by a parser of a structure containing it as a field. Instead, the custom data structure should implement ParseMetaFlatUnnamed::parse_meta_flat_unnamed, ParseMetaFlatNamed::parse_meta_flat_named, or ParseMetaRest::parse_meta_rest. Then, the implementation of this function should delegate to one of those methods.

The default implementation directly calls Self::parse_meta_item.

fn parse_meta_item_flag(_span: Span) -> Result<Self>

Parses an empty flag value.

Only called when parsing a struct with named fields, with no value given after the field name. _span is the span of the token following the name.

The default implementation returns an error.

fn parse_meta_item_named( input: ParseStream<'_>, _name: &str, span: Span, ) -> Result<Self>

Parses the item following a name.

Only called when parsing a struct with named fields. Normally implementations will not need to override this method.

The default implementation simply calls parse_helpers::parse_named_meta_item.

fn missing_meta_item(name: &str, span: Span) -> Result<Self>

Fallback for when a required item is missing.

Only called when a required (non-default) field was omitted from a parsed attribute. Implementations on types that implement Default will most likely want to return Ok(Default::default()) here.

The default implementation returns an error.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.