deluxe_core

Trait ParseMetaRest

pub trait ParseMetaRest: Sized {
    // Required method
    fn parse_meta_rest<'s, S: Borrow<ParseBuffer<'s>>>(
        inputs: &[S],
        exclude: &[&str],
    ) -> Result<Self>;
}

Parses a meta item for a structure with named fields that consumes all fields.

Should be implemented on map types. Implementations are provided for the common map types in std.

Required Methods

fn parse_meta_rest<'s, S: Borrow<ParseBuffer<'s>>>( inputs: &[S], exclude: &[&str], ) -> Result<Self>

Parse the item from a group of inline named contexts.

Fields with names in exclude should be should be skipped with crate::parse_helpers::skip_meta_item.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.