deluxe_core::validations

Function all_or_none

pub fn all_or_none<'t, I, O>(attrs: I, prefix: &str, errors: &Errors)
where
    I: IntoIterator<Item = &'t (&'static str, O)>,
    I::IntoIter: Clone,
    O: Into<Option<&'t dyn Spanned>> + Clone + ?Sized + 't,

Appends an error if some attributes are present, but not all of them.

attrs should provide an iterator of tuples containing the field name, and an Option possibly containing the field value. If at least one is Some and at least one is None, an error will be appended to errors, with prefix prepended onto the names.