deluxe_core::validations

Function only_one

pub fn only_one<'t, I, O>(attrs: I, prefix: &str, errors: &Errors)
where
    I: IntoIterator<Item = &'t (&'static str, O)>,
    I::IntoIter: Clone,
    O: Into<Option<&'t dyn Spanned>> + Clone + ?Sized + 't,

Appends an error if more than one given attribute is present.

attrs should provide an iterator of tuples containing the field name, and an Option possibly containing the field value. If two or more are Some, an error will be appended to errors, with prefix prepended onto the names.