deluxe_core
Module with

Custom parsing helpers for #[deluxe(with = ...)].

Modules

any_path

Helpers for parsing a path allowing any keywords as identifiers, and containing no path arguments.

from_str

Helpers for parsing any type that implements std::str::FromStr.

from_str_default

Helpers for parsing any type that implements std::str::FromStr and Default.

identity

Helpers for parsing any type that implements ParseMetaItem as itself.

maybe_quoted

Helpers for parsing any type that implements ParseMetaItem possibly parsed out of a quoted string first.

mod_path

Helpers for parsing a module path using syn::Path::parse_mod_style.

quoted

Helpers for parsing any type that implements ParseMetaItem parsed out of a quoted string first.

syn

Helpers for parsing any type that implements syn::parse::Parse.

syn_quoted

Helpers for parsing any type that implements syn::parse::Parse parsed out of a quoted string first.