deluxe_core::parse_helpers

Trait ParseDelimited

pub trait ParseDelimited: Sealed {
    // Required method
    fn parse_delimited(input: ParseStream<'_>) -> Result<ParseBuffer<'_>>;

    // Provided methods
    fn parse_delimited_with<T, F: FnOnce(ParseStream<'_>) -> Result<T>>(
        input: ParseStream<'_>,
        func: F,
    ) -> Result<T> { ... }
    fn parse_delimited_meta_item<T: ParseMetaItem>(
        input: ParseStream<'_>,
        mode: ParseMode,
    ) -> Result<T> { ... }
}

Helpers for parsing tokens inside a delimiter.

This trait is only implemeted for the syn delimiter token types.

Required Methods

fn parse_delimited(input: ParseStream<'_>) -> Result<ParseBuffer<'_>>

Parses a pair of delimiter tokens, and returns a ParseBuffer for the tokens inside the delimiter.

Provided Methods

fn parse_delimited_with<T, F: FnOnce(ParseStream<'_>) -> Result<T>>( input: ParseStream<'_>, func: F, ) -> Result<T>

Parse a stream surrounded by a delimiter. The inner stream is allowed to contain a trailing comma.

fn parse_delimited_meta_item<T: ParseMetaItem>( input: ParseStream<'_>, mode: ParseMode, ) -> Result<T>

Parse a ParseMetaItem surrounded by a delimiter. The inner stream is allowed to contain a trailing comma.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.